requirements.txt
python-ldap @ http://asil-sv-au1/Repository/python_ldap-3.3.1-cp38-cp38-win_amd64.whl
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
cached-property==1.5.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.7
django-auth-ldap==2.2.0
django-cors-headers==3.3.0
django-extensions==2.2.9
django-js-asset==1.2.2
django-mptt==0.11.0
django-mptt-admin==2.0.2
django-netfields==1.2.2
djangorestframework==3.11.0
djangorestframework-recursive==0.1.2
isodate==0.6.0
itypes==1.2.0
Jinja2==2.11.2
ldap3==2.9
lxml==4.6.2
MarkupSafe==1.1.1
multi-key-dict==2.0.3
pbr==5.5.1
psycopg2==2.8.5
pure-python-adb==0.3.0.dev0
pyasn1-modules==0.2.8
python-jenkins==1.7.0
requests-file==1.5.1
sqlparse==0.4.1
suds-jurko==0.6
uritemplate==3.0.1

django-scheduler==0.9.4
